move auth models to their own file
